requirements.txt
fastapi[standard]
tensorflow-cpu==2.11.0
joblib
pillow
scikit-learn==1.1.3
numpy==1.24.3